# feat(explain): quote module values